Move GitHubRelease commands to nb.exe (#114)
* Refactor GitHubRelease project to output as a library; remove Program.cs and update Visual Studio Code version in configuration files

* Refactor GitHub release commands for improved error handling

Updated `CreateRelease` and `DownloadAsset` methods in the `Command` class to return `Task<ResultHelper>` for better success and error reporting. Enhanced the `Cli` class with new parameters for repository name, branch name, and asset file name. Added a package reference for `ntools-launcher` in `GitHubRelease.csproj`. Modified the `Program` class to support asynchronous command execution. Updated build scripts in `common.targets` and `nbuild.targets` to utilize the new command-line interface for creating releases and pre-releases.

* Fix CLI command list and add release management commands

* Add GitHub release handling to CLI application

Implemented new command types for GitHub releases:
`create_pre_release`, `create_release`, and `download_release`.
Added `ValidateReleaseOptions` method to ensure required
parameters are provided. Introduced `ValidateRepo` method
to check repository format and accessibility via the
GitHub API. Enhanced error handling for invalid inputs
and HTTP request exceptions.

* Update CLI and documentation to clarify path usage for git_clone and release commands

* Update documentation to reflect changes in Nb.exe command options and clarify path usage for git_clone and release commands

* Update release CLI validation logic

* Updte CLI Tests in NBuildTests; Update Documentation

* Refactor Command class and update usage in Program.cs

Significantly redefined the `Command` class in `Command.cs`, adding detailed XML documentation for methods `CreateRelease`, `DownloadAsset`, and `UploadAsset`. Updated `using` directives to include necessary namespaces while removing obsolete ones.

In `Program.cs`, simplified method calls to `Command.CreateRelease` and `Command.DownloadAsset` by removing the namespace prefix, aligning with the new structure of the `Command` class.

* Rename release management commands in Cli class

Updated command types in the `Cli` class from `create_release`, `create_pre_release`, and `download_release` to `release_create`, `pre_release_create`, and `release_download`. Adjusted command definitions, execution logic in `Program.cs`, and test cases in `CliTests.cs` to reflect these changes. Documentation in `nbuild.md` and command descriptions have been updated for consistency. The `nbuild.targets` file was also modified to align with the new command structure.

* Update documentation, remove obsolete code, and add features

Refactor documentation comments in `ReleaseService.cs` for clarity and completeness, including parameter and return type details. Remove the obsolete method `CheckTokenPermissionsObsolsete`. Introduce a new method `VerifyAssetId` to check asset ID existence in releases. Modify `Cli.cs` to change command type from `download` to `release_download` for better path validation. Overall, these changes enhance code readability and adherence to coding standards.

* Enhance documentation and refactor methods

Added XML documentation comments to several methods in
`CommitService.cs`, `ContributorService.cs`, `ReleaseFormatter.cs`,
and `ReleaseService.cs` for improved clarity. Removed the
unnecessary `GetShaFromBranch` method and renamed
`IsStagingTag` and `IsProductionTag` to `IsStageTag` and
`IsProdTag` for consistency. Overall, these changes improve
code readability and streamline the codebase.

* Update CLI command descriptions for clarity and consistency

* Clarify error messages for required options in CLI commands

* Update CLI command descriptions

* Update nbuild command usage in docs

---------
