+++ Nbuild/Nbuild.csproj
@@ -48,6 +48,7 @@
 
 	</ItemGroup>
  <ItemGroup>
+    <ProjectReference Include="..\GitHubRelease\GitHubRelease.csproj" />
     <ProjectReference Include="..\NbuildTasks\NbuildTasks.csproj" />
   </ItemGroup>
