+++ NbuildTasks/NbuildTasks.csproj
@@ -26,7 +26,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Build.Utilities.Core" Version="17.13.9" />	  
+    <PackageReference Include="Microsoft.Build.Utilities.Core" Version="17.14.8" />	  
     <PackageReference Include="ntools-launcher" Version="1.7.0" />
   </ItemGroup>
